Split code in Calendar component
