+++ src/components/Taskbar/components/Plans/components/Calendar/Calendar.module.css
@@ -1,91 +1,5 @@
 .root {
-  --daysSpacing: calc(var(--spacing) / 5);
-
   padding: var(--spacing) calc(var(--spacing) * 2) calc(var(--spacing) * 2);
   border-bottom: 1px solid var(--plansBorderColor);
   font-size: 1.5rem;
-}
-
-.header {
-  display: flex;
-  justify-content: space-between;
-  padding-bottom: calc(var(--spacing) * 2);
-}
-
-.headerElement {
-  flex-basis: 40%;
-}
-
-.dateInformation {
-  display: flex;
-  gap: 5%;
-  cursor: pointer;
-}
-
-.dateInformation:hover {
-  filter: brightness(130%);
-}
-
-.month {
-  text-transform: lowercase;
-}
-
-.arrowsContainer {
-  display: flex;
-  justify-content: flex-end;
-  gap: 10%;
-}
-
-.arrow {
-  width: 20px;
-  height: 20px;
-  cursor: pointer;
-}
-
-.weekDays {
-  display: flex;
-  padding-bottom: var(--spacing);
-  text-transform: lowercase;
-  text-align: center;
-}
-
-.weekDay {
-  flex-basis: calc(100% / 7);
-}
-
-.days {
-  display: grid;
-  grid-template-columns: repeat(7, 1fr);
-  grid-template-rows: repeat(5, 1fr);
-  gap: var(--daysSpacing);
-}
-
-.day {
-  line-height: 35px;
-  border: 2px solid transparent;
-  text-align: center;
-}
-
-.day:hover {
-  border-color: var(--grey);
-}
-
-.PREVIOUS_MONTH_DAY,
-.NEXT_MONTH_DAY {
-  color: rgba(255, 255, 255, 0.5);
-}
-
-.TODAY {
-  padding: var(--daysSpacing);
-  border: 2px solid var(--windowsGreen);
-}
-
-.TODAY:hover {
-  border-color: var(--windowsGreenOnHover);
-}
-
-.TODAY .dayNumber {
-  background-color: var(--windowsGreen);
-  color: var(--white);
-  font-weight: bold;
 }
