+++ src/components/Taskbar/components/Plans/components/Calendar/components/DaysList/DaysList.module.css
@@ -1,0 +1,51 @@
+.root {
+  --daysSpacing: calc(var(--spacing) / 5);
+}
+
+.weekDays {
+  display: flex;
+  padding-bottom: var(--spacing);
+  text-transform: lowercase;
+  text-align: center;
+}
+
+.weekDay {
+  flex-basis: calc(100% / 7);
+}
+
+.days {
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
+  grid-template-rows: repeat(5, 1fr);
+  gap: var(--daysSpacing);
+}
+
+.day {
+  line-height: 35px;
+  border: 2px solid transparent;
+  text-align: center;
+}
+
+.day:hover {
+  border-color: var(--grey);
+}
+
+.PREVIOUS_MONTH_DAY,
+.NEXT_MONTH_DAY {
+  color: rgba(255, 255, 255, 0.5);
+}
+
+.TODAY {
+  padding: var(--daysSpacing);
+  border: 2px solid var(--windowsGreen);
+}
+
+.TODAY:hover {
+  border-color: var(--windowsGreenOnHover);
+}
+
+.TODAY .dayNumber {
+  background-color: var(--windowsGreen);
+  color: var(--white);
+  font-weight: bold;
+}
